pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aristotle-virtue-graph"
version = "0.1.0"
description = "Evidence-first knowledge graph for Aristotle's Nicomachean Ethics Book II"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "beautifulsoup4>=4.12,<5",
  "httpx>=0.27,<1",
  "jsonschema>=4.23,<5",
  "lxml>=5,<6",
  "networkx>=3.3,<4",
  "pydantic>=2.7,<3",
  "typer>=0.12,<1",
  "PyYAML>=6.0,<7",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.10,<2",
  "pytest>=8,<9",
  "ruff>=0.5,<1",
]
viewer = [
  "pyvis>=0.3.2,<1",
  "streamlit>=1.44,<2",
]

[project.scripts]
aristotle-graph = "aristotle_graph.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["B", "E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.13"
strict = true
mypy_path = "src"
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = ["bs4", "bs4.*", "jsonschema", "lxml", "networkx", "pyvis.*", "yaml"]
ignore_missing_imports = true
